Conflicts:
	src/declarative/graphicsitems/qdeclarativelistview.cpp

+++ src/declarative/graphicsitems/qdeclarativelistview_p.h
@@ -216,7 +216,7 @@ public:
     enum PositionMode { Beginning, Center, End, Visible, Contain };
 
     Q_INVOKABLE void positionViewAtIndex(int index, int mode);
-    Q_INVOKABLE int indexAt(int x, int y) const;
+    Q_INVOKABLE int indexAt(qreal x, qreal y) const;
     Q_INVOKABLE Q_REVISION(1) void positionViewAtBeginning();
     Q_INVOKABLE Q_REVISION(1) void positionViewAtEnd();
